+++ modules/Auth/SignIn/InputField.tsx
@@ -7,7 +7,7 @@ import {
   FormField,
 } from '@/components/ui/form';
 import {Input} from '@/components/ui/input';
-import {SignInSchemaType} from './validation/SignIn';
+import {SignInSchemaType} from './validator';
 import {Control} from 'react-hook-form';
 
 type Props = {
